# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyMP Backend API
  version: 2.0.0
  description: Canonical API v2 for launchers, dashboards, and SkyMP game servers.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v2/launcher/manifest:
    get:
      tags: [Launcher]
      operationId: getLauncherManifest
      responses:
        '200': { $ref: '#/components/responses/Collection' }
        '503': { $ref: '#/components/responses/JsonError' }
...
  /api/v2/auth/discord/start:
    get:
      tags: [Authentication]
      operationId: startLauncherDiscordLogin
      parameters:
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        '302': { description: Redirect to Discord }
        '400': { $ref: '#/components/responses/JsonError' }
  /api/v2/auth/discord/callback:
    get:
      tags: [Authentication]
      operationId: completeLauncherDiscordLogin
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: Browser completion page }
        '400': { description: Invalid or expired OAuth flow }
  /api/v2/auth/discord/status:
    get:
      tags: [Authentication]
      operationId: getLauncherDiscordLoginStatus
      parameters:
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
        '401': { $ref: '#/components/responses/JsonError' }
...
  /api/v2/admin/auth/session:
    get:
      tags: [Administration, Authentication]
      operationId: getDashboardSession
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
        '401': { $ref: '#/components/responses/JsonError' }
  /api/v2/admin/auth/logout:
    post:
      tags: [Administration, Authentication]
      operationId: logoutDashboardSession
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/admin/players:
    get:
      tags: [Administration]
      operationId: listPlayers
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/Collection' }
        '401': { $ref: '#/components/responses/JsonError' }
    post:
      tags: [Administration]
      operationId: createPlayer
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '201': { $ref: '#/components/responses/JsonSuccess' }
...
  /api/v2/admin/role-permissions:
    get:
      tags: [Administration]
      operationId: listRolePermissions
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/admin/role-permissions/{roleId}:
    parameters:
      - { name: roleId, in: path, required: true, schema: { type: string } }
    put:
      tags: [Administration]
      operationId: updateRolePermissions
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
    delete:
      tags: [Administration]
      operationId: deleteRolePermissions
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
  /api/v2/admin/proxy/{path}:
    parameters:
      - { name: path, in: path, required: true, schema: { type: string } }
    get:
      tags: [Administration]
      operationId: proxyAdminGet
      security: [{ dashboardBearer: [] }]
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
    post:
      tags: [Administration]
      operationId: proxyAdminPost
      security: [{ dashboardBearer: [] }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
  /api/v2/game-servers/{key}/profiles/{id}/access:
    get:
      tags: [Game servers]
      operationId: checkProfileAccess
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
components:
  securitySchemes:
    dashboardBearer:
      type: http
      scheme: bearer
  parameters:
    ServerKey:
      name: key
      in: path
      required: true
      schema: { type: string }
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    JsonSuccess:
      description: Successful JSON response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Collection:
      description: Collection response
      content:
        application/json:
          schema:
            type: object
            required: [items, total]
            properties:
              items: { type: array, items: {} }
              total: { type: integer, minimum: 0 }
    JsonError:
      description: Structured API v2 error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}